feat: support Gitlab sections (jared.allard/codeowners!1)
Adds support for sections as defined in the Gitlab documentation[^1].

Right now, the only support exposed for sections is Rules inherting
default sections. Eventually, support may be added for
grouping/organizing them by section.

An edge case not handled right now is Gitlab's handling of duplicate
sections. These are treated a seperate sections. If we end up needing to
support that, it will be added later.

[^1]: https://docs.gitlab.com/ee/user/project/codeowners/#organize-code-owners-by-putting-them-into-sections